requirements.txt
numpy 
rembg 
Flask
Pillow 
SQLAlchemy
onnxruntime
scikit-learn
opencv-python 
psycopg2-binary